[Auth] init oauth_nctu inside package
